tidy the code
